PULL_REQUEST_TEMPLATE: request only issue number
For some time now, GitHub is showing cards when we hover over an issue
number which gives us a quick summary of the issue. The issue title is
shown in the card among other things.

So, I believe it's now moot to ask the contributors to include the issue
title in the pull request description. It's better if we remove this
unnecessary burden from them.